_scripts/email/requirements.txt
PyGithub
PyYAML
boto3
beautifulsoup4
